renamed PhoenixAbstractDrive and implemented AbstractDrive interface
